delete the single server view
