+++ server/src/main/java/org/apache/druid/client/ServerInventoryViewProvider.java
@@ -27,7 +27,6 @@ import com.google.inject.Provider;
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type", defaultImpl = BatchServerInventoryViewProvider.class)
 @JsonSubTypes(value = {
-    @JsonSubTypes.Type(name = "legacy", value = SingleServerInventoryProvider.class),
     @JsonSubTypes.Type(name = "batch", value = BatchServerInventoryViewProvider.class),
     @JsonSubTypes.Type(name = "http", value = HttpServerInventoryViewProvider.class),
 })
